requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
-e git+ssh://git@github.com/Edward-HartleyICL/fyp-robotGPT.git@b3871060b6cd689c1b1b9ab061c6d971fe93ff87#egg=fyp_package
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12>=8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.16.1
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.6
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyzmq==25.1.2
requests==2.31.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
# triton>=2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
